requirements.txt
flask
pymongo
werkzeug
groq
easyocr
PyPDF2
pillow